requirements.txt
uvicorn==0.34.3
pydantic-settings==2.10.1
fastapi==0.115.13
sqlalchemy==2.0.41
sqladmin==0.20.1
asyncpg==0.30.0
dnspython==2.7.0
email-validator==2.2.0
alembic==1.16.2
python-jose==3.5.0
passlib==1.7.4
redis==6.2.0
itsdangerous==2.2.0